Сделал отдельный модуль под работу с комментами
