+++ js/rendering-full.js
@@ -2,6 +2,7 @@
 
 import {objects, picturesContainer} from './rendering-mini.js';
 import {renderComments} from './rendering-comments.js';
+import {createComments} from './create-comments.js';
 
 const bigPicture = document.querySelector('.big-picture');
 
@@ -16,66 +17,7 @@ function onMiniatureClick (evt) {
     bigPicture.querySelector('.social__likes').textContent = objects[id].likes;
     bigPicture.querySelector('.comments-count').textContent = objects[id].comments.length;
     renderComments(id);
-
-// подсчет комментариев
-    const commentContainer = document.querySelector('.social__comments');
-    const commentsLength = commentContainer.children.length;
-    bigPicture.querySelector('.visible-comments-count').textContent = commentsLength;
-
-    const commentsLoader = bigPicture.querySelector('.comments-loader');
-    const checkLoaderStatus = () => {
-      if (commentsLength <= 5) {
-        commentsLoader.classList.add('hidden');
-      } else {
-        commentsLoader.classList.remove('hidden');
-      }
-    };
-
-    checkLoaderStatus();
-
-
-    function showComments () {
-      console.log('длина комментов' + commentsLength)
-      let commentsVisibleLength = commentsLength;
-      if (commentsLength > 5) {
-        commentsVisibleLength = 5;
-      }
-      console.log('отображаемых сообщений: ' + commentsVisibleLength)
-      let commentsHiddenLength = commentsLength - commentsVisibleLength;
-      console.log('спрятанных сообщений: ' + commentsHiddenLength)
-
-      if (commentsHiddenLength <= 5) {
-        return function () {
-          for (let i = 0; i <= commentsVisibleLength + commentsHiddenLength - 1; i++) {
-            commentContainer.children[i].classList.remove('hidden');
-
-          }
-
-
-        };
-
-      }
-
-      if (commentsHiddenLength > 5) { // другой случай если остаток больше 5
-        // counter = Math.floor(commentsHiddenLength / 5);
-        // commentsHiddenLength = commentsHiddenLength % 5;
-        return function () {
-          for (let i = 0; i <= commentsVisibleLength - 1 + 5; i++) {
-            commentContainer.children[i].classList.remove('hidden');
-          }
-          commentsVisibleLength += 5;
-          return commentsVisibleLength;
-
-        };
-      }
-
-    // if (commentsHiddenLength === 0) {commentsLoader.classList.add('hidden')} // почему не работает???????? В ЦЕЛОМ НАДО ПОНЯТЬ КАК ФУНКЦИИ СОХРАНЯТЬ ДАННЫЕ МЕЖДУ ИТЕРАЦИЯМИ
-    }
-
-    const onLoaderClick = showComments();
-
-    commentsLoader.addEventListener('click', onLoaderClick);
-
+    createComments();
     document.querySelector('body').classList.add('modal-open');
   }
 }
